# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "UGM Proceeding API",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/authors/": {
            "get": {
                "tags": [
                    "authors"
                ],
                "summary": "Get all authors",
                "description": "Return collection of authors",
                "operationId": "getAllProceedings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "name values that need to be considered for filter",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
...
        "/books/": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get all books",
                "description": "Return collection of books",
                "operationId": "getAllProceedings",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/books/{bookId}": {
            "get": {
                "tags": [
                    "books"
                ],
                "summary": "Get book by ID",
                "description": "Return single record of book",
                "operationId": "getProceedingById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of book to return",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/books": {
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "Create a new Book",
                "description": "",
                "operationId": "createBook",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Book object that needs to be added",
                        "required": true,
                        "schema": {
                            "properties": {
                                "title": {
                                    "type": "string",
                                    "example": "Menuju Negara Maju: Apakah Indonesia bergerak ke arah yang benar?"
                                },
                                "description": {
                                    "type": "string",
                                    "example": "Indonesia termasuk dalam negara 20 besar perekonomian dunia yakni pendapatan Domestik Bruto terbesar ke-16 menggunakan standar current US$ dan negara dengan total PDB terbesat ke-10 menurut perhitungan alternatif menggunakan Purchasing Power Parity (PPP) (Worldbank, 2013). Indonesia juga telah berhasil keluar dari kelompok negara-negara berpendapatan bawah menjadi negara berpendapatan menengah. Namu, sebagian pakar mengatakan kemungkinan Indonesia masuk dalam jebakan pendapatan menengah (middle income trap). Buku ini bertujuan untuk memberikan gambaran tentang perkembangan capaian indikator sosial ekonomi Indonesia untuk melakukan lompatan atau catch up untuk mengejar negara-negara maju."
                                },
                                "category": {
                                    "type": "string",
                                    "example": "Economy"
                                },
                                "edition": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "pages": {
                                    "type": "integer",
                                    "example": 151
                                },
                                "publication_year": {
                                    "type": "integer",
                                    "example": 2015
                                },
                                "publisher": {
                                    "type": "string",
                                    "example": "UGM Press"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/books/{bookId}/author": {
            "post": {
                "tags": [
                    "books"
                ],
                "summary": "Add an author to book",
                "description": "",
                "operationId": "addBookAuthor",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "bookId",
                        "in": "path",
                        "description": "ID of book to return",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Book object that needs to be added",
                        "required": true,
                        "schema": {
                            "properties": {
                                "user_email": {
                                    "type": "string",
                                    "example": "ugm.proceeding@mail.com"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "405": {
                        "description": "Invalid input"
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
        "/proceedings/": {
            "get": {
                "tags": [
                    "proceedings"
                ],
                "summary": "Get all proceedings",
                "description": "Return collection of proceedings",
                "operationId": "getAllProceedings",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "keyword",
                        "in": "query",
                        "description": "keyword values that need to be considered for filter",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Name of the proceeding that need to be considered for filter",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "alias",
                        "in": "query",
                        "description": "Alias of the proceeding that need to be considered for filter. Ex: ICTA",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "subject",
                        "in": "query",
                        "description": "subject id of the proceeding that need to be considered for filter. Ex: 3",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "date",
                        "in": "query",
                        "description": "Date of the conference ex: 2017-12",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "you can sort the data based on id, name, alias, and date. Don't forget to put the direction after the field. ex: name.asc, id.desc",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status of the proceeding",
                        "required": false,
                        "type": "string",
                        "default": "all",
                        "enum": [
                            "published",
                            "draft",
                            "trashed",
                            "all"
                        ]
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/proceedings/{proceedingsId}": {
            "get": {
                "tags": [
                    "proceedings"
                ],
                "summary": "Find proceeding by Id",
                "description": "Returns a single proceeding with articles",
                "operationId": "getProceedingById",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "proceedingsId",
                        "in": "path",
                        "description": "ID of proceeding to return",
                        "required": true,
                        "type": "integer",
                        "format": "int64"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Proceeding not found"
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "Bearer": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}